Fix instructions and gulpfile.js.sample
Because of this commit 20ad2bfb8aa0267eb59867dd6d8f57013e311c67,
joomla-gulp expects to be in a subdirectory, not directly in the project
root. That's why the instruction is changed to:

    npm install --save-dev joomla-gulp

`--save-dev` has no effect unless a `package.json` file exists already,
so the `npm init` instruction needs to be moved up.

The copy instructions need to be updated because of the change in
layout.

The `gulpfile.js.sample` needs to updated as well.